rename: domRender 함수 utils로 이동
